requirements.txt
Flask
python-dotenv
langchain
huggingface_hub
werkzeug
langchain_community